requirements.txt
langchain
langchain-chroma
langchain-huggingface
langchain-community
sentence-transformers
chromadb
pypdf
openai
python-dotenv
